# Reduce memory consumption