--- a/example_test.go
+++ b/example_test.go
@@ -13,15 +13,15 @@ import (
 func ExampleWriteAll_stdout() {
 	entries := []sitemap.UrlEntry{
 		{
-			Loc:     "http://example.com/",
+			Loc:     "http://goiguide.com/",
 			LastMod: time.Date(2025, time.November, 2, 11, 34, 58, 123, time.UTC),
 		},
 		{
-			Loc: "http://example.com/test/",
+			Loc: "http://goiguide.com/test/",
 			Images: []string{
-				"http://example.com/test/1.jpg",
-				"http://example.com/test/2.jpg",
-				"http://example.com/test/3.jpg",
+				"http://goiguide.com/test/1.jpg",
+				"http://goiguide.com/test/2.jpg",
+				"http://goiguide.com/test/3.jpg",
 			},
 		},
 	}
@@ -36,19 +36,19 @@ func ExampleWriteAll_stdout() {
 	// <?xml version="1.0" encoding="UTF-8"?>
 	// <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">
 	//   <url>
-	//     <loc>http://example.com/</loc>
+	//     <loc>http://goiguide.com/</loc>
 	//     <lastmod>2025-11-02T11:34:58Z</lastmod>
 	//   </url>
 	//   <url>
-	//     <loc>http://example.com/test/</loc>
+	//     <loc>http://goiguide.com/test/</loc>
 	//     <image:image>
-	//       <image:loc>http://example.com/test/1.jpg</image:loc>
+	//       <image:loc>http://goiguide.com/test/1.jpg</image:loc>
 	//     </image:image>
 	//     <image:image>
-	//       <image:loc>http://example.com/test/2.jpg</image:loc>
+	//       <image:loc>http://goiguide.com/test/2.jpg</image:loc>
 	//     </image:image>
 	//     <image:image>
-	//       <image:loc>http://example.com/test/3.jpg</image:loc>
+	//       <image:loc>http://goiguide.com/test/3.jpg</image:loc>
 	//     </image:image>
 	//   </url>
 	// </urlset>
@@ -58,7 +58,7 @@ func ExampleWriteAll_stdout() {
 	// <?xml version="1.0" encoding="UTF-8"?>
 	// <sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
 	//   <url>
-	//     <loc>https://example.com/sitemap-0.xml</loc>
+	//     <loc>https://goiguide.com/sitemap-0.xml</loc>
 	//   </url>
 	// </sitemapindex>
 }
@@ -66,15 +66,15 @@ func ExampleWriteAll_stdout() {
 func ExampleWriteAll_buffers() {
 	entries := []sitemap.UrlEntry{
 		{
-			Loc:     "http://example.com/",
+			Loc:     "http://goiguide.com/",
 			LastMod: time.Date(2025, time.November, 2, 11, 34, 58, 123, time.UTC),
 		},
 		{
-			Loc: "http://example.com/test/",
+			Loc: "http://goiguide.com/test/",
 			Images: []string{
-				"http://example.com/test/1.jpg",
-				"http://example.com/test/2.jpg",
-				"http://example.com/test/3.jpg",
+				"http://goiguide.com/test/1.jpg",
+				"http://goiguide.com/test/2.jpg",
+				"http://goiguide.com/test/3.jpg",
 			},
 		},
 	}
@@ -98,19 +98,19 @@ func ExampleWriteAll_buffers() {
 	// <?xml version="1.0" encoding="UTF-8"?>
 	// <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">
 	//   <url>
-	//     <loc>http://example.com/</loc>
+	//     <loc>http://goiguide.com/</loc>
 	//     <lastmod>2025-11-02T11:34:58Z</lastmod>
 	//   </url>
 	//   <url>
-	//     <loc>http://example.com/test/</loc>
+	//     <loc>http://goiguide.com/test/</loc>
 	//     <image:image>
-	//       <image:loc>http://example.com/test/1.jpg</image:loc>
+	//       <image:loc>http://goiguide.com/test/1.jpg</image:loc>
 	//     </image:image>
 	//     <image:image>
-	//       <image:loc>http://example.com/test/2.jpg</image:loc>
+	//       <image:loc>http://goiguide.com/test/2.jpg</image:loc>
 	//     </image:image>
 	//     <image:image>
-	//       <image:loc>http://example.com/test/3.jpg</image:loc>
+	//       <image:loc>http://goiguide.com/test/3.jpg</image:loc>
 	//     </image:image>
 	//   </url>
 	// </urlset>
@@ -120,7 +120,50 @@ func ExampleWriteAll_buffers() {
 	// <?xml version="1.0" encoding="UTF-8"?>
 	// <sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
 	//   <url>
-	//     <loc>https://example.com/sitemap-0.xml</loc>
+	//     <loc>https://goiguide.com/sitemap-0.xml</loc>
+	//   </url>
+	// </sitemapindex>
+}
+
+func ExampleWriteAll_dynamicInput() {
+	err := sitemap.WriteAll(&stdoutOutput{}, &dynamicInput{Length: 3})
+	if err != nil {
+		fmt.Printf("Error: %s\n", err)
+	}
+	// Output:
+	// ::: Urlset 0
+
+	// <?xml version="1.0" encoding="UTF-8"?>
+	// <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">
+	//   <url>
+	//     <loc>https://goiguide.com/entry-00000</loc>
+	//     <lastmod>2020-10-31T11:00:00Z</lastmod>
+	//     <image:image>
+	//       <image:loc>https://goiguide.com/entry-00000/thumb.png</image:loc>
+	//     </image:image>
+	//   </url>
+	//   <url>
+	//     <loc>https://goiguide.com/entry-00001</loc>
+	//     <lastmod>2020-11-01T11:00:00Z</lastmod>
+	//     <image:image>
+	//       <image:loc>https://goiguide.com/entry-00001/thumb.png</image:loc>
+	//     </image:image>
+	//   </url>
+	//   <url>
+	//     <loc>https://goiguide.com/entry-00002</loc>
+	//     <lastmod>2020-11-02T11:00:00Z</lastmod>
+	//     <image:image>
+	//       <image:loc>https://goiguide.com/entry-00002/thumb.png</image:loc>
+	//     </image:image>
+	//   </url>
+	// </urlset>
+
+	// ::: Index
+
+	// <?xml version="1.0" encoding="UTF-8"?>
+	// <sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+	//   <url>
+	//     <loc>https://goiguide.com/sitemap-00.xml</loc>
 	//   </url>
 	// </sitemapindex>
 }
@@ -156,19 +199,42 @@ func (o *bufferOutput) Urlset() io.Writer {
 
 type arrayInput struct {
 	Arr     []sitemap.UrlEntry
-	NextIdx int
+	nextIdx int
 }
 
 func (a arrayInput) HasNext() bool {
-	return a.NextIdx < len(a.Arr)
+	return a.nextIdx < len(a.Arr)
 }
 
 func (a *arrayInput) Next() *sitemap.UrlEntry {
-	idx := a.NextIdx
-	a.NextIdx++
+	idx := a.nextIdx
+	a.nextIdx++
 	return &a.Arr[idx]
 }
 
 func (a *arrayInput) GetUrlsetUrl(n int) string {
-	return fmt.Sprintf("https://example.com/sitemap-%d.xml", n)
+	return fmt.Sprintf("https://goiguide.com/sitemap-%d.xml", n)
+}
+
+type dynamicInput struct {
+	Length  int
+	nextIdx int
+	entry   sitemap.UrlEntry
+}
+
+func (d dynamicInput) HasNext() bool {
+	return d.nextIdx < d.Length
+}
+
+func (d *dynamicInput) Next() *sitemap.UrlEntry {
+	idx := d.nextIdx
+	d.nextIdx++
+	d.entry.Loc = fmt.Sprintf("https://goiguide.com/entry-%05d", idx)
+	d.entry.LastMod = time.Date(2020, time.November, idx, 11, 0, 0, 0, time.UTC)
+	d.entry.Images = []string{fmt.Sprintf("https://goiguide.com/entry-%05d/thumb.png", idx)}
+	return &d.entry
+}
+
+func (d dynamicInput) GetUrlsetUrl(n int) string {
+	return fmt.Sprintf("https://goiguide.com/sitemap-%02d.xml", n)
 }