--- a/sitemap.go
+++ b/sitemap.go
@@ -7,19 +7,6 @@ import (
 	"time"
 )
 
-type Url struct {
-	XMLName xml.Name `xml:"url"`
-	Loc     string   `xml:"loc"`
-	LastMod string   `xml:"lastmod,omitempty"`
-	Images  []Image  `xml:"image:image,omitempty"`
-	//      Priority   float32   `xml:"priority"`
-	//      ChangeFreq string    `xml:"changefreq,omitempty"`
-}
-
-type Image struct {
-	Loc string `xml:"image:loc"`
-}
-
 var minDate = time.Date(2000, 1, 1, 0, 0, 0, 0, time.UTC)
 
 func t2str(t time.Time) string {
@@ -36,119 +23,141 @@ func t2str(t time.Time) string {
 // by o.Index().
 // The function aborts if any unexpected error occurs when writing.
 func WriteAll(o Output, in Input) error {
+	var s sitemapWriter
 	var nfiles int
 	for {
 		nfiles++
-		err := writeUrlsetFile(o.Urlset(), in)
+		err := s.writeUrlsetFile(o.Urlset(), in)
 		if err != nil && !errors.Is(err, errMaxCapReached{}) {
 			return err
 		}
 
 		if err == nil {
-			return writeIndexFile(o.Index(), in, nfiles)
+			return s.writeIndexFile(o.Index(), in, nfiles)
 		}
 	}
 }
 
+type sitemapWriter struct{}
+
 // writeIndexFile writes Sitemap index file for N files.
-func writeIndexFile(w io.Writer, in Input, nfiles int) error {
-	start := xml.StartElement{
-		Name: xml.Name{Local: "sitemapindex"},
-		Attr: []xml.Attr{
-			{
-				Name:  xml.Name{Local: "xmlns"},
-				Value: "http://www.sitemaps.org/schemas/sitemap/0.9",
-			},
-		},
-	}
+func (s *sitemapWriter) writeIndexFile(w io.Writer, in Input, nfiles int) error {
+	abortWriter := abortWriter{underlying: w}
 
-	if _, err := io.WriteString(w, xml.Header); err != nil {
-		return err
-	}
-
-	e := xml.NewEncoder(w)
-	e.Indent("", "  ")
-	if err := e.EncodeToken(start); err != nil {
-		return err
-	}
-
-	sitemap := Url{}
+	abortWriter.Write(indexHeader)
 	for i := 0; i < nfiles; i++ {
-		sitemap.Loc = in.GetUrlsetUrl(i)
-		if err := e.Encode(sitemap); err != nil {
-			return err
-		}
+		s.writeXmlUrlLoc(&abortWriter, in.GetUrlsetUrl(i))
 	}
+	abortWriter.Write(indexFooter)
 
-	if err := e.EncodeToken(start.End()); err != nil {
-		return err
-	}
-
-	return e.Flush()
+	return abortWriter.firstErr
 }
 
 // writeUrlsetFile writes a single Sitemap Urlset file for the first 50K entries
 // in the given input.
-func writeUrlsetFile(w io.Writer, in Input) (retErr error) {
-	start := xml.StartElement{
-		Name: xml.Name{Local: "urlset"},
-		Attr: []xml.Attr{
-			{
-				Name:  xml.Name{Local: "xmlns"},
-				Value: "http://www.sitemaps.org/schemas/sitemap/0.9",
-			},
-			{
-				Name:  xml.Name{Local: "xmlns:image"},
-				Value: "http://www.google.com/schemas/sitemap-image/1.1",
-			},
-		},
-	}
+func (s *sitemapWriter) writeUrlsetFile(w io.Writer, in Input) error {
+	abortWriter := abortWriter{underlying: w}
+	var capErr error
 
-	if _, err := io.WriteString(w, xml.Header); err != nil {
-		return err
-	}
-
-	e := xml.NewEncoder(w)
-	e.Indent("", "  ")
-	if err := e.EncodeToken(start); err != nil {
-		return err
-	}
-
-	url := Url{}
-	image := Image{}
-	var count int
-	for in.HasNext() {
+	abortWriter.Write(urlsetHeader)
+	for count := 0; in.HasNext(); count++ {
 		if count >= maxSitemapCap {
-			retErr = errMaxCapReached{}
+			capErr = errMaxCapReached{}
 			break
 		}
 
-		entry := in.Next()
+		s.writeXmlUrlEntry(&abortWriter, in.Next())
+	}
+	abortWriter.Write(urlsetFooter)
 
-		url.Loc = entry.GetLoc()
-		url.LastMod = t2str(entry.GetLastMod())
-		url.Images = []Image{}
-		for _, imageUrl := range entry.GetImages() {
-			image.Loc = imageUrl
-			url.Images = append(url.Images, image)
+	if abortWriter.firstErr != nil {
+		return abortWriter.firstErr
+	}
+
+	return capErr
+}
+
+func (s *sitemapWriter) writeXmlUrlEntry(w io.Writer, e UrlEntry) {
+	loc := e.GetLoc()
+	lastMod := t2str(e.GetLastMod())
+	images := e.GetImages()
+	if lastMod == "" && len(images) == 0 {
+		// fast path
+		s.writeXmlUrlLoc(w, loc)
+		return
+	}
+
+	w.Write(tagUrlOpen)
+	w.Write(tagLocOpen)
+	s.writeXmlString(w, loc)
+	w.Write(tagLocClose)
+	if lastMod != "" {
+		w.Write(tagLastmodOpen)
+		s.writeXmlString(w, lastMod)
+		w.Write(tagLastmodClose)
+	}
+	if len(images) > 0 {
+		for i := range images {
+			w.Write(tagImageOpen)
+			s.writeXmlString(w, images[i])
+			w.Write(tagImageClose)
 		}
+	}
+	w.Write(tagUrlClose)
+}
 
-		if err := e.Encode(url); err != nil {
-			return err
-		}
+func (s *sitemapWriter) writeXmlUrlLoc(w io.Writer, loc string) {
+	w.Write(tagUrlOpen)
+	w.Write(tagLocOpen)
+	s.writeXmlString(w, loc)
+	w.Write(tagLocClose)
+	w.Write(tagUrlClose)
+}
 
-		count++
+func (s *sitemapWriter) writeXmlString(w io.Writer, str string) {
+	xml.EscapeText(w, []byte(str))
+}
+
+// Below are constant strings converted to byte slices ahead of time
+// to avoid run-time allocations caused by string to byte slice conversions.
+var (
+	indexHeader = []byte(xml.Header +
+		`<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">` +
+		"\n",
+	)
+	indexFooter = []byte("</sitemapindex>")
+
+	urlsetHeader = []byte(xml.Header +
+		`<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">` +
+		"\n",
+	)
+	urlsetFooter = []byte(`</urlset>`)
+
+	tagUrlOpen      = []byte("  <url>\n")
+	tagUrlClose     = []byte("  </url>\n")
+	tagLocOpen      = []byte("    <loc>")
+	tagLocClose     = []byte("</loc>\n")
+	tagLastmodOpen  = []byte("    <lastmod>")
+	tagLastmodClose = []byte("</lastmod>\n")
+	tagImageOpen    = []byte("    <image:image>\n      <image:loc>")
+	tagImageClose   = []byte("</image:loc>\n    </image:image>\n")
+)
+
+type abortWriter struct {
+	underlying io.Writer
+	firstErr   error
+}
+
+func (w *abortWriter) Write(p []byte) (n int, err error) {
+	if w.firstErr != nil {
+		return 0, w.firstErr
 	}
 
-	if err := e.EncodeToken(start.End()); err != nil {
-		return err
+	n, err = w.underlying.Write(p)
+	if err != nil {
+		w.firstErr = err
 	}
-
-	if err := e.Flush(); err != nil {
-		return err
-	}
-
-	return retErr
+	return
 }
 
 const (