This PR reduces the memory allocation performed by this library, and makes it independent of the input length, `O(1)`.

This was possible due to the following changes:
- Convert all XML code to constant strings (https://github.com/PlanitarInc/go-sitemap/pull/14/commits/26ee6421a425b6e9b3eb68f6d8ecc996499962b8). All the XML code (tags and attributes) is known ahead of time so there is no need to render it using `xml.Encoder`.
- Convert all constant string to byte slices ahead of time (https://github.com/PlanitarInc/go-sitemap/pull/14/commits/26ee6421a425b6e9b3eb68f6d8ecc996499962b8). When rendering an XML file to `io.Writer` we need to write a slice of bytes rather than a string. Converting a string to a byte slice results in an allocation because GO strings are immutable.
- Use a temporary buffer for rendering dynamic string values to XML (https://github.com/PlanitarInc/go-sitemap/pull/14/commits/fa7179a4989b190b1ab885c363e25210b244b37b). Again, this is needed to remove the allocation made when converting a string to a slice of bytes.
- Use a temporary buffer for rendering time values to XML (https://github.com/PlanitarInc/go-sitemap/pull/14/commits/d4da84f85b0a91a589e08274c2c26042b275e309). This approach is a bit tricky. We rely on 2 facts: (a) the final serialized time string will have the same length as the `time.RFC3339` format string, (b) `time.AppendFormat()` is trying to reuse the passed slice of bytes for rendering a string. The approach is to pre-allocate enough space in the temporary buffer and to pass the byte slice of this buffer to `AppendFormat()`. In case any of the stated 2 assumptions is broken, the code will continue work as expected but it won't be "alloc-free".
- Use struct for `UrlEntry` type (https://github.com/PlanitarInc/go-sitemap/pull/14/commits/53dd41cbad2c165bd1620689c069298227584d9b). This is needed to remove the allocation made for "boxing" a concrete entry type (defined by an application) into library's `UrlEntry` interface upon every call of `Input.Next()`.

By applying these changes we've made the library almost "alloc-free". The allocations now depend on `UrlEntry.Loc` https://github.com/PlanitarInc/go-sitemap/blob/53dd41cbad2c165bd1620689c069298227584d9b/interfaces.go#L14-L15 rather on the length of input. Side note: a carefully designed input can cause this library to allocate for every rendered entry but this should not really happen in practice.

On the way, this PR introduces `abortWriter` to simplify the code. This writer type is a wrapper for a writer, all it is doing is pretty much recording the first writing error. 


### Benchmarks

The outcome of these changes is that the library is running at least 60% faster.

```
$ go test -run '^$' -bench . -benchmem
goos: darwin
goarch: amd64
pkg: github.com/PlanitarInc/go-sitemap
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkWriteAll/1-12                      	 1323837	       920.8 ns/op	     128 B/op	       3 allocs/op
BenchmarkWriteAll/10-12                     	  174440	      6726 ns/op	     128 B/op	       3 allocs/op
BenchmarkWriteAll/100-12                    	   18532	     65129 ns/op	     128 B/op	       3 allocs/op
BenchmarkWriteAll/1000-12                   	    1855	    640370 ns/op	     128 B/op	       3 allocs/op
BenchmarkWriteAll/10000-12                  	     168	   7395119 ns/op	     128 B/op	       3 allocs/op
BenchmarkWriteAll/100000-12                 	      16	  65864465 ns/op	     160 B/op	       4 allocs/op
BenchmarkWriteAll/1000000-12                	       2	 747052939 ns/op	     736 B/op	      22 allocs/op
BenchmarkWriteUrlset/1-12                   	 1398517	       849.3 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/10-12                  	  181693	      6681 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/100-12                 	   16810	     73855 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/1000-12                	    1707	    681426 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/10000-12               	     168	   7014104 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/100000-12              	      32	  35761613 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteUrlset/1000000-12             	      36	  32433611 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/1-12                    	 9478348	       127.2 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/10-12                   	 1214608	       894.0 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/100-12                  	  140778	      8384 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/1000-12                 	   14365	     82711 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/10000-12                	    1311	    859726 ns/op	      32 B/op	       1 allocs/op
BenchmarkWriteIndex/100000-12               	     141	   8356007 ns/op	      32 B/op	       1 allocs/op
BenchmarkSitemapWriter_writeXmlString/short-12         	31168033	        37.79 ns/op	       0 B/op	       0 allocs/op
BenchmarkSitemapWriter_writeXmlString/long-12          	 5605268	       209.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkSitemapWriter_writeXmlTime/utc-12             	 5869622	       205.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkSitemapWriter_writeXmlTime/custom-12          	 5462307	       221.6 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/PlanitarInc/go-sitemap	39.145s
```

Diff with master:

```
$ benchcmp old.bench new.bench
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                           old ns/op      new ns/op     delta
BenchmarkWriteAll/1-12              5811           921           -84.15%
BenchmarkWriteAll/10-12             23392          6726          -71.25%
BenchmarkWriteAll/100-12            198466         65129         -67.18%
BenchmarkWriteAll/1000-12           1952371        640370        -67.20%
BenchmarkWriteAll/10000-12          19451444       7395119       -61.98%
BenchmarkWriteAll/100000-12         194844889      65864465      -66.20%
BenchmarkWriteAll/1000000-12        1938904848     747052939     -61.47%
BenchmarkWriteUrlset/1-12           3711           849           -77.11%
BenchmarkWriteUrlset/10-12          21358          6681          -68.72%
BenchmarkWriteUrlset/100-12         195381         73855         -62.20%
BenchmarkWriteUrlset/1000-12        1943058        681426        -64.93%
BenchmarkWriteUrlset/10000-12       21040814       7014104       -66.66%
BenchmarkWriteUrlset/100000-12      101490329      35761613      -64.76%
BenchmarkWriteUrlset/1000000-12     96949476       32433611      -66.55%
BenchmarkWriteIndex/1-12            1912           127           -93.35%
BenchmarkWriteIndex/10-12           6545           894           -86.34%
BenchmarkWriteIndex/100-12          51513          8384          -83.72%
BenchmarkWriteIndex/1000-12         513240         82711         -83.88%
BenchmarkWriteIndex/10000-12        5323145        859726        -83.85%
BenchmarkWriteIndex/100000-12       60147152       8356007       -86.11%

benchmark                           old allocs     new allocs     delta
BenchmarkWriteAll/1-12              25             3              -88.00%
BenchmarkWriteAll/10-12             61             3              -95.08%
BenchmarkWriteAll/100-12            421            3              -99.29%
BenchmarkWriteAll/1000-12           4021           3              -99.93%
BenchmarkWriteAll/10000-12          40021          3              -99.99%
BenchmarkWriteAll/100000-12         400032         4              -100.00%
BenchmarkWriteAll/1000000-12        4000235        22             -100.00%
BenchmarkWriteUrlset/1-12           14             1              -92.86%
BenchmarkWriteUrlset/10-12          50             1              -98.00%
BenchmarkWriteUrlset/100-12         410            1              -99.76%
BenchmarkWriteUrlset/1000-12        4010           1              -99.98%
BenchmarkWriteUrlset/10000-12       40010          1              -100.00%
BenchmarkWriteUrlset/100000-12      200010         1              -100.00%
BenchmarkWriteUrlset/1000000-12     200010         1              -100.00%
BenchmarkWriteIndex/1-12            11             1              -90.91%
BenchmarkWriteIndex/10-12           20             1              -95.00%
BenchmarkWriteIndex/100-12          110            1              -99.09%
BenchmarkWriteIndex/1000-12         1010           1              -99.90%
BenchmarkWriteIndex/10000-12        10010          1              -99.99%
BenchmarkWriteIndex/100000-12       100010         1              -100.00%

benchmark                           old bytes     new bytes     delta
BenchmarkWriteAll/1-12              9720          128           -98.68%
BenchmarkWriteAll/10-12             11520         128           -98.89%
BenchmarkWriteAll/100-12            29520         128           -99.57%
BenchmarkWriteAll/1000-12           209520        128           -99.94%
BenchmarkWriteAll/10000-12          2009520       128           -99.99%
BenchmarkWriteAll/100000-12         20014352      160           -100.00%
BenchmarkWriteAll/1000000-12        200101720     736           -100.00%
BenchmarkWriteUrlset/1-12           4936          32            -99.35%
BenchmarkWriteUrlset/10-12          6736          32            -99.52%
BenchmarkWriteUrlset/100-12         24736         32            -99.87%
BenchmarkWriteUrlset/1000-12        204736        32            -99.98%
BenchmarkWriteUrlset/10000-12       2004737       32            -100.00%
BenchmarkWriteUrlset/100000-12      10004752      32            -100.00%
BenchmarkWriteUrlset/1000000-12     10004736      32            -100.00%
BenchmarkWriteIndex/1-12            4784          32            -99.33%
BenchmarkWriteIndex/10-12           5648          32            -99.43%
BenchmarkWriteIndex/100-12          14288         32            -99.78%
BenchmarkWriteIndex/1000-12         100688        32            -99.97%
BenchmarkWriteIndex/10000-12        964689        32            -100.00%
BenchmarkWriteIndex/100000-12       9604693       32            -100.00%
```